requirements.txt
django==2.0.1
gitpython==2.1.8
httplib2==0.10.3
